@@ -3,26 +3,32 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "lint": "eslint --ext .js,.vue --ignore-path .gitignore --fix src/"
+  },
+  "engines": {
+    "npm": ">=8.0.0",
+    "node": ">=18.0.0"
   },
   "dependencies": {
     "@mediapipe/face_detection": "^0.4.1646425229",
     "@tensorflow-models/face-detection": "^1.0.2",
-    "@tensorflow/tfjs-backend-webgl": "^4.12.0",
-    "@tensorflow/tfjs-converter": "^4.12.0",
+    "@tensorflow/tfjs-backend-webgl": "^4.15.0",
+    "@tensorflow/tfjs-converter": "^4.15.0",
     "@tensorflow/tfjs-core": "^4.12.0",
-    "core-js": "^3.6.5",
-    "vue": "^2.6.11"
+    "@vitejs/plugin-vue": "^5.0.2",
+    "core-js": "^3.35.0",
+    "eslint-config-prettier": "^9.1.0",
+    "prettier": "^3.1.1",
+    "vite": "^5.0.10",
+    "vite-plugin-eslint": "^1.8.1",
+    "vue": "^3.4.3"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.12",
-    "@vue/cli-plugin-eslint": "~4.5.12",
-    "@vue/cli-service": "~4.5.12",
     "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11"
+    "eslint": "^8.56.0",
+    "eslint-plugin-vue": "^9.19.2"
   }
 }
