@@ -1,8 +1,6 @@
-import Vue from 'vue'
-import App from './App.vue'
+import {createApp, h} from 'vue';
+import App from './App.vue';
 
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+createApp({
+  render: () => h(App),
+}).mount('#app');
